Update   - Test Sound Click For TitlePage  - Change SoundEffect,BgSound   Integer to double

+++ Models/Setting.cs
@@ -5,9 +5,9 @@ namespace BeyondHana.Models
     {
        
         [ObservableProperty]
-        private int backgroundmusicpercent;
+        private double backgroundmusicpercent;
         [ObservableProperty]
-        private int soundeffectpercent;
+        private double soundeffectpercent;
         [ObservableProperty]
         private int textsize;
     }
